+++ src/backend/mailbox_manager/tests/api/mail_domain/test_api_mail_domains_retrieve.py
@@ -2,7 +2,11 @@
 Tests for MailDomains API endpoint in People's mailbox manager app. Focus on "retrieve" action.
 """
 
+import json
+import re
+
 import pytest
+import responses
 from rest_framework import status
 from rest_framework.test import APIClient
 
@@ -14,6 +18,7 @@ from mailbox_manager.tests.fixtures import dimail as dimail_fixtures
 pytestmark = pytest.mark.django_db
 
 
+@responses.activate
 def test_api_mail_domains__retrieve_anonymous():
     """Anonymous users should not be allowed to retrieve a domain."""
 
@@ -24,8 +29,11 @@ def test_api_mail_domains__retrieve_anonymous():
     assert response.json() == {
         "detail": "Authentication credentials were not provided."
     }
+    # Verify no calls were made to dimail API
+    assert len(responses.calls) == 0
 
 
+@responses.activate
 def test_api_domains__retrieve_non_existing():
     """
     Authenticated users should have an explicit error when trying to retrive
@@ -39,8 +47,11 @@ def test_api_domains__retrieve_non_existing():
     )
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json() == {"detail": "Not found."}
+    # Verify no calls were made to dimail API
+    assert len(responses.calls) == 0
 
 
+@responses.activate
 def test_api_mail_domains__retrieve_authenticated_unrelated():
     """
     Authenticated users should not be allowed to retrieve a domain
@@ -60,6 +71,7 @@ def test_api_mail_domains__retrieve_authenticated_unrelated():
     assert response.json() == {"detail": "No MailDomain matches the given query."}
 
 
+@responses.activate
 def test_api_mail_domains__retrieve_authenticated_related():
     """
     Authenticated users should be allowed to retrieve a domain
@@ -91,9 +103,11 @@ def test_api_mail_domains__retrieve_authenticated_related():
         "support_email": domain.support_email,
         "last_check_details": None,
         "action_required_details": {},
+        "expected_config": None,
     }
 
 
+@responses.activate
 def test_api_mail_domains__retrieve_authenticated_related_with_action_required():
     """
     Authenticated users should be allowed to retrieve a domain
@@ -113,7 +127,6 @@ def test_api_mail_domains__retrieve_authenticated_related_with_action_required()
     response = client.get(
         f"/api/v1.0/mail-domains/{domain.slug}/",
     )
-
     assert response.status_code == status.HTTP_200_OK
     assert response.json() == {
         "id": str(domain.id),
@@ -130,9 +143,11 @@ def test_api_mail_domains__retrieve_authenticated_related_with_action_required()
             "mx": "Je veux que le MX du domaine soit mx.ox.numerique.gouv.fr., "
             "or je trouve example-fr.mail.protection.outlook.com.",
         },
+        "expected_config": None,
     }
 
 
+@responses.activate
 def test_api_mail_domains__retrieve_authenticated_related_with_ok_status():
     """
     Authenticated users should be allowed to retrieve a domain
@@ -165,4 +180,43 @@ def test_api_mail_domains__retrieve_authenticated_related_with_ok_status():
         "support_email": domain.support_email,
         "last_check_details": dimail_fixtures.CHECK_DOMAIN_OK,
         "action_required_details": {},
+        "expected_config": None,
+    }
+
+
+@responses.activate
+def test_api_mail_domains__retrieve_authenticated_related_with_failed_status():
+    """
+    Authenticated users should be allowed to retrieve a domain
+    to which they have access and which has failed status.
+    """
+    user = core_factories.UserFactory()
+
+    client = APIClient()
+    client.force_login(user)
+
+    domain = factories.MailDomainFactory(
+        status=enums.MailDomainStatusChoices.FAILED,
+        last_check_details=dimail_fixtures.CHECK_DOMAIN_BROKEN_INTERNAL,
+    )
+    factories.MailDomainAccessFactory(domain=domain, user=user)
+
+    response = client.get(
+        f"/api/v1.0/mail-domains/{domain.slug}/",
+    )
+
+    assert response.status_code == status.HTTP_200_OK
+    assert response.json() == {
+        "id": str(domain.id),
+        "name": domain.name,
+        "slug": domain.slug,
+        "status": domain.status,
+        "created_at": domain.created_at.isoformat().replace("+00:00", "Z"),
+        "updated_at": domain.updated_at.isoformat().replace("+00:00", "Z"),
+        "abilities": domain.get_abilities(user),
+        "count_mailboxes": 0,
+        "support_email": domain.support_email,
+        "last_check_details": dimail_fixtures.CHECK_DOMAIN_BROKEN_INTERNAL,
+        "action_required_details": {},
+        "expected_config": None,
     }